Extensible non regression tests
