+++ src/main/groovy/com/lesfurets/jenkins/helpers/cps/BaseRegressionTestCPS.groovy
@@ -1,17 +1,7 @@
 package com.lesfurets.jenkins.helpers.cps
 
-import static com.lesfurets.jenkins.helpers.MethodCall.callArgsToString
-import static org.assertj.core.api.Assertions.assertThat
-import static org.assertj.core.api.Assertions.fail
+import com.lesfurets.jenkins.helpers.RegressionTest
 
-import java.util.function.Function
-import java.util.stream.Collectors
-
-import org.assertj.core.api.SoftAssertions
-
-import com.lesfurets.jenkins.helpers.*
-
-abstract class BaseRegressionTestCPS extends BasePipelineTestCPS {
-
+abstract class BaseRegressionTestCPS extends BasePipelineTestCPS implements RegressionTest {
 
 }
